refactor(playback): extract handle_pipeline_event to sibling module
playback/mod.rs was at 993/1000 lines — any future edit would break the
1000-line airuleset cap. handle_pipeline_event is a natural extraction
boundary: 180 lines, self-contained, dispatches on PipelineEvent.

Moved into crates/sp-server/src/playback/handle_pipeline_event.rs as an
extension impl block on PlaybackEngine. mod.rs declares it via plain
`mod handle_pipeline_event;`. The orphaned `use title::OBS_TITLE_SOURCE;`
in mod.rs is removed (only the moved fn used it; `title::push_title` and
`title::get_video_title_info` callers in mod.rs use the module path
directly).

playback/mod.rs is now 811 lines (was 993, +189 headroom).

Closes #95
